callable: Implement a hook system
Define the concept of ``hook`` and ``catch``. A hook is a function which when it
is executed also executes other functions catching it.

This can be useful in an event based context a hook function ``on_event`` is
defined and other function catching this event are implicitly called as part of
the event.

Signed-off-by: Jimmy Thrasibule <dev@jimmy.lt>