task/abc: Implement a task abstraction
A task is a piece of code which must be executed in a controlled fashion. A task
wraps execution of its payload under different error control mechanisms. At each
phase of the task execution flow, an event is triggered.

A class implementing a task, can catch any of the triggered events to control
the task execution.

Signed-off-by: Jimmy Thrasibule <dev@jimmy.lt>